Split options into separate library
